% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{aero_resistance}
\alias{aero_resistance}
\title{Aerodynamic Resistance}
\usage{
aero_resistance(wind_speed, height, crop_height)
}
\arguments{
\item{wind_speed}{Wind speed (m/s)}

\item{height}{Measured height of wind speed (m)}

\item{crop_height}{Crop height (m)}
}
\value{
Aerodynamic resistance (day/m)
}
\description{
Aerodynamic Resistance
}
\examples{
aero_resistance(2, 3, 1.4)
}
