% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{calc_aero_resistance}
\alias{calc_aero_resistance}
\title{Calculate Aerodynamic Resistance}
\usage{
calc_aero_resistance(wind_speed, height, crop_height)
}
\arguments{
\item{wind_speed}{Wind speed (m/s)}

\item{height}{Wind measurement height (m)}

\item{crop_height}{Crop height (m)}
}
\value{
Aerodynamic resistance (day/m)
}
\description{
Calculate Aerodynamic Resistance
}
\examples{
cropsyst:::calc_aero_resistance(rep(2, 30), 3, 1.4)
}
