% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{calc_air_entry_potential}
\alias{calc_air_entry_potential}
\title{Calculate air entry potential}
\usage{
calc_air_entry_potential(sand_pct, clay_pct, b_value, saturation_wc)
}
\arguments{
\item{sand_pct}{Percentage of sand in soil layer}

\item{clay_pct}{Percentage of clay in soil layer}

\item{b_value}{Campbell "B" value}

\item{saturation_wc}{Saturation water content (m3/m3)}
}
\value{
Air entry potential (J/kg)
}
\description{
Calculate air entry potential
}
\examples{
cropsyst:::calc_air_entry_potential(92, 3, -4.045, 0.326)
}
