% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{calc_bulk_density}
\alias{calc_bulk_density}
\title{Calculate bulk density}
\usage{
calc_bulk_density(saturation_wc)
}
\arguments{
\item{saturation_wc}{Saturation Water Content (m3/m3)}
}
\value{
Bulk density (kg/m3)
}
\description{
Calculate bulk density
}
\examples{
cropsyst:::calc_bulk_density(0.326)
}
