% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_cover.R
\name{calc_canopy_cover}
\alias{calc_canopy_cover}
\title{Calculate Canopy Cover}
\usage{
calc_canopy_cover(ndvi)
}
\arguments{
\item{ndvi}{a time series of ndvi values from planting to harvest date}
}
\value{
canopy cover
}
\description{
Calculate Canopy Cover
}
\examples{
calc_canopy_cover(runif(100))
}
