% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{calc_daytime_vpd}
\alias{calc_daytime_vpd}
\title{Calculate Daytime Vapor Pressure Deficit}
\usage{
calc_daytime_vpd(tmax, rhmin)
}
\arguments{
\item{tmax}{Daily maximum temperature (C)}

\item{rhmin}{Daily minimum relative humidity}
}
\value{
Vapor Pressure Deficit (kPa)
}
\description{
Calculate Daytime Vapor Pressure Deficit
}
\examples{
cropsyst:::calc_daytime_vpd(30, 30)
}
