% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{calc_potential_radiation}
\alias{calc_potential_radiation}
\title{Calculate potential radiation}
\usage{
calc_potential_radiation(lat, doy)
}
\arguments{
\item{lat}{latitude (degrees)}

\item{doy}{day of year}
}
\value{
Potential Radiation (MJ/m2/day)
}
\description{
Calculate potential radiation
}
\examples{
cropsyst:::calc_potential_radiation(45, 233)
}
