% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{calc_sat_vapor_pressure}
\alias{calc_sat_vapor_pressure}
\title{Calculate saturation vapor pressure}
\usage{
calc_sat_vapor_pressure(temp)
}
\arguments{
\item{temp}{Air temperature (°C)}
}
\value{
Saturation Vapor Pressure (kPa)
}
\description{
Calculate saturation vapor pressure
}
\examples{
cropsyst:::calc_sat_vapor_pressure(32)
}
