% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{calc_saturated_hydraulic_conductivity}
\alias{calc_saturated_hydraulic_conductivity}
\title{Calculate saturated hydraulic conductivity}
\usage{
calc_saturated_hydraulic_conductivity(sand_pct, clay_pct, saturation_wc)
}
\arguments{
\item{sand_pct}{Percentage of sand in soil layer}

\item{clay_pct}{Percentage of clay in soil layer}

\item{saturation_wc}{Saturation water content (m3/m3)}
}
\value{
Saturated hydraulic conductivity (kg s/m3)
}
\description{
Calculate saturated hydraulic conductivity
}
\examples{
cropsyst:::calc_saturated_hydraulic_conductivity(92, 3, 0.00341)
}
