% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{calc_vapor_pressure}
\alias{calc_vapor_pressure}
\title{Calculate vapor pressure}
\usage{
calc_vapor_pressure(es_tmax, es_tmin, rhmax, rhmin)
}
\arguments{
\item{es_tmax}{Saturation Vapor Pressure at Maximum Temperature (°C)}

\item{es_tmin}{Saturation Vapor Pressure at Minimum Temperature (°C)}

\item{rhmax}{Maximum relative humidity (\%)}

\item{rhmin}{Minimum relative humidity (\%)}
}
\value{
Vapor Pressure (kPA)
}
\description{
Calculate vapor pressure
}
\examples{
cropsyst:::calc_vapor_pressure(5, 3, 100, 40)
}
