% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{calc_water_potential_field_capacity}
\alias{calc_water_potential_field_capacity}
\title{Calculate water potential at field capacity}
\usage{
calc_water_potential_field_capacity(silt_pct, clay_pct)
}
\arguments{
\item{silt_pct}{Percentage of sand in soil layer}

\item{clay_pct}{Percentage of clay in soil layer}
}
\value{
Water potential at field capacity (m3/m3)
}
\description{
Calculate water potential at field capacity
}
\examples{
cropsyst:::calc_water_potential_field_capacity(92, 3)
}
