% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cs_add_crop}
\alias{cs_add_crop}
\title{Add crop parameters to cropsyst object}
\usage{
cs_add_crop(
  cs,
  planting_date,
  season_end_date,
  midseason_et_crop_coef,
  tree_fruit,
  fruit_harvest_date,
  max_crop_height,
  max_root_depth,
  perrenial,
  max_water_uptake,
  transpiration_use_eff,
  C3
)
}
\arguments{
\item{cs}{cropsyst object}

\item{planting_date}{Planting date}

\item{season_end_date}{Season end date}

\item{midseason_et_crop_coef}{mid season ET crop coefficient}

\item{tree_fruit}{logical TRUE if tree fruit crop}

\item{fruit_harvest_date}{harvest date of tree fruit}

\item{max_crop_height}{max crop height (m)}

\item{max_root_depth}{max root depth (m)}

\item{perrenial}{logical TRUE if perennial}

\item{max_water_uptake}{max water uptake}

\item{transpiration_use_eff}{transpiration use efficiency}

\item{C3}{C3 = TRUE c4 = FALSE}
}
\value{
cropsyst object with added crop parameters
}
\description{
Add crop parameters to cropsyst object
}
