% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cs_add_irrigation}
\alias{cs_add_irrigation}
\title{Add irrigation data to cropsyst object}
\usage{
cs_add_irrigation(
  cs,
  irrigation,
  irrigated_fraction,
  max_allowable_depletion,
  min_leaf_water_potential,
  max_system_capacity,
  dae_first_irrigation,
  overhead_irrigation = FALSE
)
}
\arguments{
\item{cs}{cropsyst object}

\item{irrigation}{irrigation data frame if missing auto irrigation is applied}

\item{irrigated_fraction}{fraction of land irrigated in simulated area}

\item{max_allowable_depletion}{Max soil water depletion which triggers irrigation}

\item{min_leaf_water_potential}{Min leaf water potential which triggers irrigation}

\item{max_system_capacity}{Max irrigation daily capacity}

\item{dae_first_irrigation}{Day after emergence of first irrigation}

\item{overhead_irrigation}{logical is irrigation applied overhead defaults to FALSE}
}
\value{
cropsyst object with added irrigation data and parameters
}
\description{
Add irrigation data to cropsyst object
}
\examples{
cs_add_irrigation(cs_setup(), cs_irrigation, 0.001)
}
