% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cs_add_soil}
\alias{cs_add_soil}
\title{Add soil data to cropsyst object}
\usage{
cs_add_soil(cs, soil, thickness_evaporative_layer)
}
\arguments{
\item{cs}{cropsyst object}

\item{soil}{soil data frame}

\item{thickness_evaporative_layer}{thickness of soil evaporative layer}
}
\value{
cropsyst object with added soil data and parameters
}
\description{
Add soil data to cropsyst object
}
\examples{
cs_add_soil(cs_setup(), cs_soil, 0.005)
}
