% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cs_add_weather}
\alias{cs_add_weather}
\title{Add weather data to cropsyst object}
\usage{
cs_add_weather(cs, weather, latitude, elevation, screening_height)
}
\arguments{
\item{cs}{cropsyst object}

\item{weather}{weather data frame}

\item{latitude}{location latitude}

\item{elevation}{elevation (m)}

\item{screening_height}{Height of wind measurement (m)}
}
\value{
cropsyst object with added weather data and parameters
}
\description{
Add weather data to cropsyst object
}
\examples{
cs_add_weather(cs_setup(), cs_weather, 45, 1000, 2)
}
