% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_canopy_cover}
\alias{cs_canopy_cover}
\title{Canopy Cover}
\format{
A data frame with 151 rows and 2 variables:
\describe{
\item{dap}{days after planting}
\item{gcc}{Green Canopy Cover}
\item{tcc}{Total Canopy Cover}
}
}
\usage{
cs_canopy_cover
}
\description{
Canopy Cover
}
\keyword{datasets}
