% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_crop}
\alias{cs_crop}
\title{Crop input paramters}
\format{
A data frame with 21 rows and 2 variables:
\describe{
\item{crop}{Crop name}
\item{midseason_et_crop_coef}{Midseason ET crop Coefficient}
\item{max_crop_height}{Max crop height (m)}
\item{max_root_depth}{Max root depth (m)}
\item{max_allowable_depletion}{Soil water depletion for no stress}
\item{transpiration_use_eff}{Transpiration-use efficiency (g/kg)}
}
}
\usage{
cs_crop
}
\description{
Crop input paramters
}
\keyword{datasets}
