% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_irrigation}
\alias{cs_irrigation}
\title{Time series of irrigation application rates}
\format{
A data frame with 21 rows and 2 variables:
\describe{
\item{dap}{day after planting}
\item{irrigation}{irrigation (mm)}
}
}
\usage{
cs_irrigation
}
\description{
Time series of irrigation application rates
}
\keyword{datasets}
