% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{cs_process_soil}
\alias{cs_process_soil}
\title{Process input soil data frame}
\usage{
cs_process_soil(soil, thickness_soil_evap_layer)
}
\arguments{
\item{soil}{a data frame with columns (horizon_thickness, sand_pct, clay_pct)
and can include intial_wc}

\item{thickness_soil_evap_layer}{thickness of evaporation layer}
}
\value{
soil data frame with added columns
saturation_wc (saturation water content),
horizon_bulk_density (bulk density),
b_value (Campbell "B" value),
air_entry_potential (air entry potential),
horizon_field_capacity (horizon field capacity),
horizon_permanent_wilting_point (horizon permanent wilting point),
layer_thickness (layer thickness), and
index (layer index)
}
\description{
Process input soil data frame
}
\examples{
cropsyst:::cs_process_soil(cs_soil, 0.005)
}
