% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{cs_process_weather}
\alias{cs_process_weather}
\title{Process Weather File}
\usage{
cs_process_weather(weather, lat, elevation, screen_height)
}
\arguments{
\item{weather}{input weather data frame}

\item{lat}{latitude}

\item{elevation}{elevelation (m)}

\item{screen_height}{screening height (m)}
}
\value{
data frame with added variables reference_et and daytime_vpd
}
\description{
Process Weather File
}
\examples{
cropsyst:::cs_process_weather(cs_weather, 45, 1000, 2)
}
