% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cs_setup}
\alias{cs_setup}
\title{Setup CropSyst}
\usage{
cs_setup(file_path = NULL)
}
\arguments{
\item{file_path}{if provided returns CropSyst input data}
}
\value{
Empty CropSyst object
}
\description{
Setup CropSyst
}
\examples{
cs_setup()
}
