% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_soil}
\alias{cs_soil}
\title{Sample soil data}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{horizon_thickness}{horizon thickness, in m}
\item{sand_pct}{percentage of sand}
\item{clay_pct}{percentage of clay}
\item{silt_pct}{percentage of silt}
\item{initial_wc}{initial water content, in m3/m3}
}
}
\usage{
cs_soil
}
\description{
A dataset containing the horizon thickness and composition of a soil cross-section
}
\keyword{datasets}
