% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_weather}
\alias{cs_weather}
\title{Sample daily weather data}
\format{
A data frame with 365 rows and 10 variables:
\describe{
\item{date}{date}
\item{precip}{daily precipitation, in mm}
\item{tmax}{maximum daily temperature, in Celcius}
\item{tmin}{minimum daily temperature, in Celcius}
\item{s_rad}{total daily solar radiation, in MJ}
\item{rh_max}{maximum daily relative humidity}
\item{rh_min}{minimum daily relative humidity}
\item{wind_speed}{wind speed, in m/s}
}
}
\usage{
cs_weather
}
\description{
A dataset containing daily weather data.
}
\keyword{datasets}
