% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{net_radiation}
\alias{net_radiation}
\title{Daily Net Radiation}
\usage{
net_radiation(pot_rad, sw_rad, ea, tmax, tmin, elevation)
}
\arguments{
\item{pot_rad}{Potential solar radiation (MJ/m2/day)}

\item{sw_rad}{Shortwave radiation (MJ/m2/day)}

\item{ea}{Daily average vapor pressure (kPa)}

\item{tmax}{Daily maximum temperature (C)}

\item{tmin}{Daily minimum temperature (C)}

\item{elevation}{elevation (m)}
}
\value{
Net Radiation (MJ/m2/day)
}
\description{
Daily Net Radiation
}
\examples{
net_radiation(900, 800, 1, 30, 10, 300)
}
