% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{potential_radiation}
\alias{potential_radiation}
\title{Potential Radiation}
\usage{
potential_radiation(lat, doy)
}
\arguments{
\item{lat}{latitude (degrees)}

\item{doy}{day of year}
}
\value{
Potential Radiation (MJ/m2/day)
}
\description{
Potential Radiation
}
\examples{
potential_radiation(45, 233)
}
