% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{sat_vapor_pressure}
\alias{sat_vapor_pressure}
\title{Saturation Vapor Pressure}
\usage{
sat_vapor_pressure(temp)
}
\arguments{
\item{temp}{Air temperature (°C)}
}
\value{
Saturation Vapor Pressure (kPa)
}
\description{
Saturation Vapor Pressure
}
\examples{
sat_vapor_pressure(32)
}
