% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{vapor_pressure}
\alias{vapor_pressure}
\title{Vapor Pressure}
\usage{
vapor_pressure(es_tmax, es_tmin, rhmax, rhmin)
}
\arguments{
\item{es_tmax}{Saturation Vapor Pressure at Maximum Temperature (°C)}

\item{es_tmin}{Saturation Vapor Pressure at Minimum Temperature (°C)}

\item{rhmax}{Maximum relative humidity (\%)}

\item{rhmin}{Minimum relative humidity (\%)}
}
\value{
Vapor Pressure (kPA)
}
\description{
Vapor Pressure
}
\examples{
vapor_pressure(5, 3, 100, 40)
}
