% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_functions.R
\name{vapor_pressure_deficit}
\alias{vapor_pressure_deficit}
\title{Vapor Pressure Deficit}
\usage{
vapor_pressure_deficit(es_tmax, es_tmin, ea)
}
\arguments{
\item{es_tmax}{Saturation Vapor Pressure at Maximum Temperature (°C)}

\item{es_tmin}{Saturation Vapor Pressure at Minimum Temperature (°C)}

\item{ea}{Vapor Pressure at Air Temperature (°C)}
}
\value{
Vapor Pressure Deficit (kPa)
}
\description{
Vapor Pressure Deficit
}
\examples{
vapor_pressure_deficit(5, 3, 4)
}
