% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydoy.R
\name{ydoy}
\alias{ydoy}
\title{Parse dates with year and day of year}
\usage{
ydoy(year, doy)
}
\arguments{
\item{year}{a numeric vector of year}

\item{doy}{a numeric vector of day of year}
}
\value{
date object
}
\description{
Parse dates with year and day of year
}
\examples{
ydoy(2016, 131)
}
